more backtraces

+++ netvr-unity/Assets/Native/IsblDynamicLibrary.cs
@@ -42,6 +42,7 @@ class IsblDynamicLibrary : IDisposable
     }
     IntPtr _library;
     readonly string _fullPath;
+    readonly bool _deleteOnDispose;
 
     public void GetDelegate<TDelegate>(string name, out TDelegate val)
     {
@@ -96,33 +97,37 @@ class IsblDynamicLibrary : IDisposable
         }
     }
 
-    public IsblDynamicLibrary(string name, string prefix)
+    public IsblDynamicLibrary(string name, string prefix, string source = null)
     {
 #if UNITY_EDITOR_WIN
-        for (int i = 0; ; ++i)
+        _deleteOnDispose = false;
+        if (source == null)
         {
-            _fullPath = prefix + name + i;
+            source = name;
+            name = source + "0";
+            _deleteOnDispose = true;
+        }
+
+        _fullPath = prefix + name;
+        var sourcePath = prefix + source;
+
+        bool same = false;
+        using (var watch = new IsblStopwatch("AreFilesIdentical:.dll"))
+        {
+            same = AreFilesIdentical(sourcePath + ".dll", _fullPath + ".dll")
+                && AreFilesIdentical(sourcePath + ".pdb", _fullPath + ".pdb");
+        }
+        if (!same)
+        {
+            // copy to new file so that original is still writeable
+            File.Copy(sourcePath + ".dll", _fullPath + ".dll", true);
             try
             {
-                using (var watch = new IsblStopwatch("AreFilesIdentical:.dll"))
-                    if (AreFilesIdentical(prefix + name + ".dll", _fullPath + ".dll") && AreFilesIdentical(prefix + name + ".pdb", _fullPath + ".pdb")) break;
-                // copy to new file so that original is still writeable
-                File.Copy(prefix + name + ".dll", _fullPath + ".dll", true);
-                try
-                {
-                    File.Copy(prefix + name + ".pdb", _fullPath + ".pdb", true);
-                }
-                catch (IOException) { }
-                break;
+                File.Copy(sourcePath + ".pdb", _fullPath + ".pdb", true);
             }
-            catch (IOException) when (i < 9)
+            catch (IOException exception)
             {
-                // retry
-            }
-            catch (IOException e)
-            {
-                Utils.LogException(e);
-                break;
+                Debug.LogWarning(exception);
             }
         }
 
@@ -130,7 +135,7 @@ class IsblDynamicLibrary : IDisposable
         Utils.Log($"Loading library from {_fullPath}.dll");
         _library = SystemLibrary.LoadLibrary(_fullPath + ".dll");
         if (_library == default) Utils.LogWarning($"Failed to load {_fullPath}.dll");
-#endif 
+#endif
     }
 
     public void Dispose()
@@ -141,18 +146,21 @@ class IsblDynamicLibrary : IDisposable
             Utils.Log("IsblDynamicLibrary.Dispose()");
             SystemLibrary.FreeLibrary(_library);
             _library = default;
-            try
+            if (_deleteOnDispose)
             {
-                File.Delete(_fullPath + ".dll");
-                File.Delete(_fullPath + ".dll.meta");
+                try
+                {
+                    File.Delete(_fullPath + ".dll");
+                    File.Delete(_fullPath + ".dll.meta");
+                }
+                catch (IOException e) { Utils.LogException(e); }
+                try
+                {
+                    File.Delete(_fullPath + ".pdb");
+                    File.Delete(_fullPath + ".pdb.meta");
+                }
+                catch (IOException) { }
             }
-            catch (IOException e) { Utils.LogException(e); }
-            try
-            {
-                File.Delete(_fullPath + ".pdb");
-                File.Delete(_fullPath + ".pdb.meta");
-            }
-            catch (IOException) { }
         }
 #endif
     }
